+++ src/pages/Employees/EmployeeForm.js
@@ -1,23 +1,7 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import { Grid } from "@material-ui/core";
 import Controls from "../../components/controls/Controls";
-import * as personAPI from "../../services/personAPI";
-import { raiseNotification } from "../../index";
 import { makeStyles } from "@material-ui/core/styles";
-
-const initialFValues = {
-  id: 0,
-  firstName: "",
-  lastName: "",
-  jobTitle: "",
-  role: "",
-  organisation: "",
-  department: "",
-  email: "",
-  phoneWork: "",
-  phoneMobile: "",
-  location: "",
-};
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -30,152 +14,55 @@ const useStyles = makeStyles((theme) => ({
 
 export default function EmployeeForm(props) {
   const classes = useStyles();
-  const [errors, setErrors] = useState({});
-  const [userProfile, setUserProfile] = useState([]);
-  const { mode } = props;
-
-  useEffect(() => {
-    const personId = props.match.params.id;
-    const getUser = async () => {
-      try {
-        const res = await personAPI.getPersonById(personId);
-        if (res.status === 200 && res.data !== null) {
-          setUserProfile(res.data);
-        } else {
-          raiseNotification(
-            "Error",
-            "Could not get user with id: " + personId,
-            "danger"
-          );
-        }
-      } catch (err) {
-        raiseNotification("API Error", "me " + err.message, "danger");
-      }
-    };
-    if (mode === "edit") {
-      getUser();
-    }
-  }, [props, mode]);
-
-  const handleInputChange = (e) => {
-    //e.preventDefault();
-    const { name, value } = e.target;
-
-    setUserProfile({
-      ...userProfile,
-      [name]: value,
-    });
-
-    validate({ [name]: value });
-  };
-
-  const validate = (fieldValues = userProfile) => {
-    let temp = { ...errors };
-    if ("firstName" in fieldValues)
-      temp.firstName = fieldValues.firstName ? "" : "This field is required.";
-    if ("lastName" in fieldValues)
-      temp.lastName = fieldValues.lastName ? "" : "This field is required.";
-    if ("email" in fieldValues)
-      temp.email = /$^|.+@.+..+/.test(fieldValues.email)
-        ? ""
-        : "Email is not valid.";
-    // if ("department" in fieldValues)
-    //   temp.department =
-    //     fieldValues.department.length !== 0 ? "" : "This field is required.";
-    // if ("jobTitle" in fieldValues)
-    //   temp.jobTitle = fieldValues.jobTitle ? "" : "This field is required.";
-
-    setErrors({
-      ...temp,
-    });
-
-    if (fieldValues === userProfile)
-      return Object.values(temp).every((x) => x === "");
-  };
-
-  const handleSubmit = async (e) => {
-    //e.preventDefault();
-    if (mode === "new") {
-      const { firstName, lastName } = userProfile;
-      if (validate()) {
-        const res = await personAPI.newPerson(userProfile);
-        if (res.status === 200 && res.data !== null) {
-          resetForm();
-
-          raiseNotification(
-            "",
-            `New person ${firstName} ${lastName} added successfully`,
-            "success"
-          );
-        }
-      }
-      props.history.push("/person/all");
-    } else {
-      const updatedPerson = userProfile;
-
-      try {
-        const res = await personAPI.updatePerson(updatedPerson);
-        if (res.status === 200 && res.data !== null) {
-          raiseNotification(
-            "",
-            `Profile '${updatedPerson.firstName} ${updatedPerson.lastName}' was updated successfully`,
-            "success"
-          );
-        } else {
-          raiseNotification(
-            "Error",
-            `Could not update user ${updatedPerson.firstName} ${updatedPerson.lastName}`,
-            "danger"
-          );
-        }
-      } catch (err) {
-        raiseNotification("API Error", err.message, "danger");
-      }
-    }
-  };
-
-  const resetForm = () => {
-    setUserProfile(initialFValues);
-    setErrors({});
-  };
+  const {
+    mode,
+    onInputChange,
+    contactProfile,
+    errors,
+    onNewContact,
+    onUpdateContact,
+    onResetForm,
+  } = props;
 
   return (
-    <form onSubmit={handleSubmit} className={classes.root} autoComplete="off">
+    <form className={classes.root} autoComplete="off">
       <Grid container>
         <Grid item xs={6}>
           <Controls.Input
             name="firstName"
             label="First name"
-            value={userProfile.firstName ? userProfile.firstName : ""}
-            onChange={handleInputChange}
+            value={contactProfile.firstName ? contactProfile.firstName : ""}
+            onChange={onInputChange}
             error={errors.firstName}
           />
           <Controls.Input
             name="lastName"
             label="Last name"
-            value={userProfile.lastName ? userProfile.lastName : ""}
-            onChange={handleInputChange}
+            value={contactProfile.lastName ? contactProfile.lastName : ""}
+            onChange={onInputChange}
             error={errors.lastName}
           />
           <Controls.Input
             name="jobTitle"
             label="Job Title"
-            value={userProfile.jobTitle ? userProfile.jobTitle : ""}
-            onChange={handleInputChange}
+            value={contactProfile.jobTitle ? contactProfile.jobTitle : ""}
+            onChange={onInputChange}
             error={errors.jobTitle}
           />
           <Controls.Input
             name="role"
             label="Role"
-            value={userProfile.role ? userProfile.role : ""}
-            onChange={handleInputChange}
+            value={contactProfile.role ? contactProfile.role : ""}
+            onChange={onInputChange}
             error={errors.role}
           />
           <Controls.Input
             name="organisation"
             label="Organisation"
-            value={userProfile.organisation ? userProfile.organisation : ""}
-            onChange={handleInputChange}
+            value={
+              contactProfile.organisation ? contactProfile.organisation : ""
+            }
+            onChange={onInputChange}
             error={errors.organisation}
           />
         </Grid>
@@ -183,45 +70,59 @@ export default function EmployeeForm(props) {
           <Controls.Input
             label="Department"
             name="department"
-            value={userProfile.department ? userProfile.department : ""}
-            onChange={handleInputChange}
+            value={contactProfile.department ? contactProfile.department : ""}
+            onChange={onInputChange}
             error={errors.department}
           />
           <Controls.Input
             label="Email"
             name="email"
-            value={userProfile.email ? userProfile.email : ""}
-            onChange={handleInputChange}
+            value={contactProfile.email ? contactProfile.email : ""}
+            onChange={onInputChange}
             error={errors.email}
           />
           <Controls.Input
             label="Location"
             name="location"
-            value={userProfile.location ? userProfile.location : ""}
-            onChange={handleInputChange}
+            value={contactProfile.location ? contactProfile.location : ""}
+            onChange={onInputChange}
             error={errors.location}
           />
           <Controls.Input
             label="Phone (work)"
             name="phoneWork"
-            value={userProfile.phoneWork ? userProfile.phoneWork : ""}
-            onChange={handleInputChange}
+            value={contactProfile.phoneWork ? contactProfile.phoneWork : ""}
+            onChange={onInputChange}
             error={errors.phoneWork}
           />
           <Controls.Input
             label="Phone (mob)"
             name="phoneMobile"
-            value={userProfile.phoneMobile ? userProfile.phoneMobile : ""}
-            onChange={handleInputChange}
+            value={contactProfile.phoneMobile ? contactProfile.phoneMobile : ""}
+            onChange={onInputChange}
             error={errors.phoneMobile}
           />
 
           <div>
+            {mode === "new" ? (
+              <Controls.Button
+                type="submit"
+                text="Create"
+                onClick={onNewContact}
+              />
+            ) : (
+              <Controls.Button
+                type="submit"
+                text="Update"
+                onClick={onUpdateContact}
+              />
+            )}
+
             <Controls.Button
-              type="submit"
-              text={mode === "new" ? "Submit" : "Update"}
+              text="Reset"
+              color="default"
+              onClick={onResetForm}
             />
-            <Controls.Button text="Reset" color="default" onClick={resetForm} />
           </div>
         </Grid>
       </Grid>
