+++ src/pages/Employees/Employees.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import {
   Table,
   Checkbox,
@@ -20,8 +20,6 @@ import FilterListIcon from "@material-ui/icons/FilterList";
 import Button from "@material-ui/core/Button";
 import DeleteIcon from "@material-ui/icons/Delete";
 import EditIcon from "@material-ui/icons/Edit";
-import { raiseNotification } from "../../index";
-import { getAllPersons, deletePerson } from "../../services/personAPI";
 import PropTypes from "prop-types";
 
 const useStyles = makeStyles((theme) => ({
@@ -130,12 +128,12 @@ EnhancedTableToolbar.propTypes = {
 function EnhancedTableHead(props) {
   const {
     classes,
-    onSelectAllClick,
     order,
     orderBy,
     numSelected,
     rowCount,
     onRequestSort,
+    onSelectAllClick,
   } = props;
   const createSortHandler = (property) => (event) => {
     onRequestSort(event, property);
@@ -189,64 +187,11 @@ EnhancedTableHead.propTypes = {
 };
 
 export default function Employees(props) {
-  const [records, setRecords] = useState([]);
-  const [selected, setSelected] = useState([]);
   const [order, setOrder] = React.useState("asc");
   const [page, setPage] = React.useState(0);
   const [orderBy, setOrderBy] = React.useState("firstName");
   const [rowsPerPage, setRowsPerPage] = React.useState(5);
-
-  useEffect(() => {
-    const getAllUsers = async () => {
-      try {
-        const res = await getAllPersons();
-        if (res.status === 200 && res.data !== null) {
-          setRecords(res.data);
-        }
-      } catch (err) {
-        raiseNotification(
-          "Error",
-          `Person could not get all persons: ${err.message}`,
-          "danger"
-        );
-      }
-    };
-    getAllUsers();
-  }, []);
-
-  const handleDelete = async () => {
-    setSelected(
-      records.filter((r) => selected.filter((sr) => sr.id === r.id).length < 1)
-    );
-
-    for (let i = 0; i < selected.length; i++) {
-      try {
-        const res = await deletePerson(selected[i]);
-        if (res.data.deletedCount === 1) {
-          raiseNotification("", "Person was deleted successfully", "success");
-          setSelected([]);
-          try {
-            const res = await getAllPersons();
-            if (res.status === 200 && res.data !== null) {
-              setRecords(res.data);
-            }
-          } catch (err) {
-            raiseNotification(
-              "Error",
-              `Person could not get all persons: ${err.message}`,
-              "danger"
-            );
-          }
-        }
-      } catch (err) {
-        raiseNotification(
-          "Error",
-          `Person could not be deleted: ${err.message}`,
-          "danger"
-        );
-      }
-    }
-  };
+  const { records, onDelete, onSelectAllClick, onRowClick, selected } = props;
 
   const classes = useStyles();
 
@@ -295,35 +240,6 @@ export default function Employees(props) {
     setOrderBy(property);
   };
 
-  const handleClick = (event, id) => {
-    const selectedIndex = selected.indexOf(id);
-    let newSelected = [];
-
-    if (selectedIndex === -1) {
-      newSelected = newSelected.concat(selected, id);
-    } else if (selectedIndex === 0) {
-      newSelected = newSelected.concat(selected.slice(1));
-    } else if (selectedIndex === selected.length - 1) {
-      newSelected = newSelected.concat(selected.slice(0, -1));
-    } else if (selectedIndex > 0) {
-      newSelected = newSelected.concat(
-        selected.slice(0, selectedIndex),
-        selected.slice(selectedIndex + 1)
-      );
-    }
-
-    setSelected(newSelected);
-  };
-
-  const handleSelectAllClick = (event) => {
-    if (event.target.checked) {
-      const newSelecteds = records.map((n) => n.id);
-      setSelected(newSelecteds);
-      return;
-    }
-    setSelected([]);
-  };
-
   const handleChangePage = (event, newPage) => {
     setPage(newPage);
   };
@@ -354,7 +270,7 @@ export default function Employees(props) {
         {/* <EmployeeForm /> */}
         <EnhancedTableToolbar
           numSelected={selected.length}
-          onDelete={handleDelete}
+          onDelete={onDelete}
         />
         {/* <Controls.Input
             label="Search Employees"
@@ -375,7 +291,7 @@ export default function Employees(props) {
             numSelected={selected.length}
             order={order}
             orderBy={orderBy}
-            onSelectAllClick={handleSelectAllClick}
+            onSelectAllClick={onSelectAllClick}
             onRequestSort={handleRequestSort}
             rowCount={records.length}
           />
@@ -388,7 +304,7 @@ export default function Employees(props) {
                   key={item.id}
                   hover
                   role="checkbox"
-                  onClick={(event) => handleClick(event, item.id)}
+                  onClick={(event) => onRowClick(event, item.id)}
                 >
                   <TableCell padding="checkbox">
                     <Checkbox
